requirements.txt
werkzeug==2.2.2
Flask==2.2.0
Flask-Admin==1.6.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.3
wtforms==2.3.3
Flask-WTF==1.2.1
marshmallow-sqlalchemy==1.0.0
email_validator==2.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
passlib==1.7.4
python-dateutil==2.9.0.post0
